+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * `QueryParameter` parameter to pass arbitrary query string parameters on the request.
 * `Top` parameter, to control how many results are returned by the API.
 * Added verbose messages to `Invoke-YTRestMethod` that show the request being made to the API.
+* `Remove-YTProject`: accept project objects, project IDs, or project short names from the pipeline.
 
 ### Changed